+++ server/db/FUN_PRO_VIEWS/13_view_form.sql
@@ -1,7 +1,8 @@
-drop view if exists view_form;
-create view view_form
-as
-select distinct cast(form.id as unsigned) form_id, sft_if.content as form_name
-from user_input ui
-left join sections form  on (ui.id_section_form = form.id)
-LEFT JOIN sections_fields_translation AS sft_if ON sft_if.id_sections = ui.id_section_form AND sft_if.id_fields = 57;
+DROP VIEW IF EXISTS view_form;
+CREATE VIEW view_form
+AS
+SELECT DISTINCT cast(form.id AS UNSIGNED) form_id, sft_if.content AS form_name, IFNULL(sft_intern.content, 0) AS internal
+FROM user_input ui
+LEFT JOIN sections form  ON (ui.id_section_form = form.id)
+LEFT JOIN sections_fields_translation AS sft_if ON sft_if.id_sections = ui.id_section_form AND sft_if.id_fields = 57
+LEFT JOIN sections_fields_translation AS sft_intern ON sft_intern.id_sections = ui.id_section_form AND sft_intern.id_fields = get_field_id('internal');