+++ server/db/FUN_PRO_VIEWS/14_view_data_tables.sql
@@ -1,10 +1,10 @@
-drop view if exists view_data_tables;
-create view view_data_tables
-as
-select 'dynamic' as type, form_id as id, form_name as orig_name, concat(form_name, '_dynamic') as table_name, CONCAT(form_id,"-","dynamic") AS form_id_plus_type
-from view_form
+DROP VIEW IF EXISTS view_data_tables;
+CREATE VIEW view_data_tables
+AS
+SELECT 'dynamic' AS `type`, form_id AS id, form_name AS orig_name, concat(form_name, '_dynamic') AS `table_name`, CONCAT(form_id,"-","dynamic") AS form_id_plus_type, internal
+FROM view_form
 
-union
+UNION
 
-select 'static' as type, id as id, name as orig_name, concat(name, '_static') as table_name, CONCAT(FLOOR(id),"-","static") AS form_id_plus_type
-from uploadTables;
+SELECT 'static' AS `type`, id AS id, `name` AS orig_name, CONCAT(`name`, '_static') AS `table_name`, CONCAT(FLOOR(id),"-","static") AS form_id_plus_type, 0  AS internal
+FROM uploadTables;